# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APZ Legal Practice Management System API
security:
  - sessionCookie: []
  - bearerSession: []
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matters/{matterId}/tasks:
    get:
      operationId: listMatterTasks
      tags: [tasks]
      summary: List tasks for a matter
      parameters:
        - name: matterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: addMatterTask
      tags: [tasks]
      summary: Create a task
      parameters:
        - name: matterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /matters/{matterId}/tasks/{id}:
    patch:
      operationId: updateMatterTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: matterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteMatterTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: matterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List the authenticated staff user's global task queue
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Global task queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task from the global queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task from the global queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # ─── CONFLICT CHECK ──────────────────────────────────────────────────────────
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [time]
      summary: List time entries
      parameters:
        - name: matterId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: billed
          in: query
          schema:
            type: boolean
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: practiceArea
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"

    post:
      operationId: createTimeEntry
      tags: [time]
      summary: Log a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryInput"
      responses:
        "201":
          description: Time entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{id}:
    patch:
      operationId: updateTimeEntry
      tags: [time]
      summary: Update a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"

    delete:
      operationId: deleteTimeEntry
      tags: [time]
      summary: Delete a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTIVITY ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
    bearerSession:
      type: http
      scheme: bearer
      description: Opaque staff session token, not a JWT.
  schemas:
...
    Task:
      type: object
      required: [id, matterId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        matterId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: ["string", "null"]
        assignedToName:
          type: ["string", "null"]
        matterReference:
          type: ["string", "null"]
        matterTitle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
        assignedToId:
          type: integer
    GlobalTaskInput:
      allOf:
        - $ref: "#/components/schemas/TaskInput"
        - type: object
          required: [matterId]
          properties:
            matterId:
              type: integer
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string

    # ─── CONFLICT ─────────────────────────────────────────────────────────────
...
    TimeEntry:
      type: object
      required: [id, matterId, description, hours, rate, total, billed, createdAt]
      properties:
        id:
          type: integer
        matterId:
          type: integer
        matterTitle:
          type: ["string", "null"]
        description:
          type: string
        hours:
          type: number
        rate:
          type: number
        total:
          type: number
        billed:
          type: boolean
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    TimeEntryInput:
      type: object
      required: [matterId, description, hours, rate]
      properties:
        matterId:
          type: integer
        description:
          type: string
        hours:
          type: number
        rate:
          type: number
        date:
          type: string
          format: date
    TimeEntryUpdate:
      type: object
      properties:
        description:
          type: string
        hours:
          type: number
        rate:
          type: number
        billed:
          type: boolean

    # ─── BILLING ──────────────────────────────────────────────────────────────
...